decorator install書き換え
