+++ app/decorators/user_decorator.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module UserDecorator
+
+  def avatar_image
+    if profile&.avatar&.attached?
+      profile.avatar
+    else 
+      'hot.jpg'
+    end  
+  end
+  
+end
