+++ app/models/article.rb
@@ -34,17 +34,5 @@ class Article < ApplicationRecord
   has_one_attached :picture
 
   belongs_to :user
-
-  def display_created_at
-    I18n.l(self.created_at, format: :long)
-  end
-
-  def author_name
-    user.user_name
-  end
-
-  def like_count
-    likes.count
-  end
   
 end
