+++ test/decorators/user_decorator_test.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+require 'test_helper'
+
+class UserDecoratorTest < ActiveSupport::TestCase
+  def setup
+    @user = User.new.extend UserDecorator
+  end
+
+  # test "the truth" do
+  #   assert true
+  # end
+end
